refactor: inline weight extraction into WhiteBoxPIM and guard the reachable failure
